Change to Score.java Change to checkHitBlock import java.util.ArrayList to Main.java Add List<Block> blocks = new ArrayList<>()to saveGame

+++ src/main/java/brickGame/Bonus.java
@@ -16,8 +16,8 @@ public class Bonus implements Serializable {
     public boolean taken = false;
 
     public Bonus(int row, int column) {
-        x = (column * (Block.getWidth())) + Block.getPaddingH() + (Block.getWidth() / 2) - 15;
-        y = (row * (Block.getHeight())) + Block.getPaddingTop() + (Block.getHeight() / 2) - 15;
+        x = (column * (Block.getWidth())) + Block.getPaddingH() + ((double) Block.getWidth() / 2) - 15;
+        y = (row * (Block.getHeight())) + Block.getPaddingTop() + ((double) Block.getHeight() / 2) - 15;
 
         draw();
     }
